refactor: move/rename file for sidecarApi
